Converted navbar.js into jsx file
